requirements.txt
PyQt5==5.15.4
Pillow==8.0.1